requirements.txt




feedparser
Flask
python-dotenv
requests
firebase-cl
schedule